requirements.txt
web3==6.0.0
pyyaml